Implement hover and go to definition providers for citations (#29)
* Show citation details on hover

* Fix markdown newline formatting issue
Replace newline with double space then a newline to actually get a newline. Double newline is a paragraph break, which we don't want.

* Implement go to definition provider
Show raw bibtex on Ctrl+hover, and go to item's location in bibtex file on click. Same with F12.

* Fix regex for getting citation keys
Was only working with the first citation key on a line

* Simplified regexp for selecting citeKeys
VSCode has an inbuilt function to make this a lot easier